Fix duplicate code + add int16 transpose support
