+++ tensorflow/lite/micro/kernels/pad.h
@@ -1,4 +1,4 @@
-/* Copyright 2022 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2025 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -20,6 +20,14 @@ limitations under the License.
 #include "tensorflow/lite/micro/kernels/kernel_util.h"
 
 namespace tflite {
+
+struct OpData {
+  PadParams params;
+  int32_t output_zero_point;
+};
+
+void* PadInit(TfLiteContext* context, const char* buffer, size_t length);
+TfLiteStatus PadPrepare(TfLiteContext* context, TfLiteNode* node);
 
 TFLMRegistration Register_PAD();
 TFLMRegistration Register_PADV2();
