+++ tensorflow/lite/micro/tools/make/ext_libs/cmsis_nn_download.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright 2024 The TensorFlow Authors. All Rights Reserved.
+# Copyright 2025 The TensorFlow Authors. All Rights Reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
